backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy>=2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
email-validator==2.1.0
openai==1.12.0
firebase-admin>=6.0.0
# openai==1.12.0 passes `proxies=` into httpx.Client; httpx 0.28 removed it.
httpx>=0.27.0,<0.28
